Small naming clean up.
